add promt test cases.
