+++ package.json
@@ -8,8 +8,8 @@
     "lint": "eslint .",
     "build": "rollup --config",
     "cov:clean": "rimraf .nyc_output && rimraf coverage",
-    "cov:unit": "node --experimental-vm-modules node_modules/jest/bin/jest.js --forceExit --colors -c jest.config.js",
-    "test": "npm-run-all cov:clean cov:unit"
+    "cov": "NODE_OPTIONS=--experimental-vm-modules jest --forceExit --colors --detectOpenHandles -c jest.config.js",
+    "test": "npm-run-all cov:clean cov"
   },
   "repository": {
     "type": "git",
@@ -43,6 +43,7 @@
     "@jest/globals": "^29.7.0",
     "@rollup/plugin-commonjs": "^26.0.1",
     "@rollup/plugin-node-resolve": "^15.2.3",
+    "cli-prompts-test": "^0.3.0",
     "eslint": "^8.57.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-prettier": "^9.1.0",
@@ -50,7 +51,6 @@
     "eslint-plugin-react-hooks": "^4.6.2",
     "eslint-plugin-react-refresh": "^0.4.7",
     "execa": "^9.2.0",
-    "fs-extra": "^11.2.0",
     "jest": "^29.7.0",
     "npm-run-all": "^4.1.5",
     "rollup": "^4.18.0",
